update frontend project structure
