+++ frontend/src/components/CriteriaWidget.tsx
@@ -1,11 +1,10 @@
-import {ReactElement} from "react";
+import { ReactElement } from "react";
 import Criteria from "../Criteria.ts";
 
-export default function CriteriaWidget(criteria:Criteria): ReactElement {
-    
-    return (
-        <>
-            <p>sup I'm the widget</p>
-        </>
-    );
+export default function CriteriaWidget(criteria: Criteria): ReactElement {
+  return (
+    <>
+      <p>sup I'm the widget</p>
+    </>
+  );
 }
