+++ frontend/src/components/RubricSideBar.tsx
@@ -1,8 +1,8 @@
 export default function RubricSideBar() {
-    return (
-        <div>
-            <h1>Group Progress</h1>
-            <p>Progress Bar</p>
-        </div>
-    );
+  return (
+    <div>
+      <h1>Group Progress</h1>
+      <p>Progress Bar</p>
+    </div>
+  );
 }
