+++ frontend/src/components/rubric-builder/CriteriaInput.tsx
@@ -1,0 +1,5 @@
+import { ReactElement } from "react";
+
+export default function CriteriaInput(): ReactElement {
+  return <></>;
+}
