+++ frontend/src/components/rubric-builder/RatingInput.tsx
@@ -1,0 +1,5 @@
+import { ReactElement } from "react";
+
+export default function RatingInput(): ReactElement {
+  return <></>;
+}
