+++ frontend/src/main.tsx
@@ -1,20 +1,20 @@
-import {StrictMode} from "react";
-import {createRoot} from "react-dom/client";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 
 import "./index.css";
-import Home from "./components/Home.tsx";
-import {Route, BrowserRouter as Router, Routes} from "react-router-dom";
-import RubricBuilder from "./components/RubricBuilder.tsx";
+import Home from "./components/views/Home.tsx";
+import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
+import RubricBuilder from "./components/views/RubricBuilder.tsx";
 
 createRoot(document.getElementById("root")!).render(
-    <StrictMode>
-        <Router>
-            <div>
-                <Routes>
-                    <Route path="/" element={<Home />} />
-                    <Route path="/rubric-builder" element={<RubricBuilder />} />
-                </Routes>
-            </div>
-        </Router>
-    </StrictMode>
+  <StrictMode>
+    <Router>
+      <div>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/rubric-builder" element={<RubricBuilder />} />
+        </Routes>
+      </div>
+    </Router>
+  </StrictMode>,
 );
